@@ -15,8 +15,10 @@
     <description>Demo project for Spring Boot</description>
 
     <properties>
-        <java.version>1.8</java.version>
-        <io-rest-assured.version>4.1.1</io-rest-assured.version>
+        <start-class>com.pw.weatherapi.WeatherapiApplication</start-class>
+        <java.version>14</java.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     </properties>
 
     <dependencies>
